requirements.txt
streamlit
rembg
opencv-python
Pillow
numpy
streamlit-drawable-canvas